OE-185 move tests to their own namespace
